+++ src/loaders/BaseDownloader.js
@@ -3,7 +3,7 @@
 import Axios from 'axios'
 import {uuid, formatPercents, randomHex, fixFileName4Windows, calcDownloadMaxConcurrency} from '../utils/LoadUtil'
 import {BaseLoader} from './BaseLoader'
-import {delay, isNetworkError, isOssUrlExpired} from '../utils/HttpUtil'
+import {isNetworkError, isStopableError, isOssUrlExpired} from '../utils/HttpUtil'
 import {formatSize, elapse} from '../utils/Formatter'
 import {getFreeDiskSize} from '../utils/FileUtil'
 
@@ -217,7 +217,7 @@ export class BaseDownloader extends BaseLoader {
       console.error(e.stack)
     }
 
-    if (isNetworkError(e)) {
+    if (isNetworkError(e) || isStopableError(e)) {
       this.stop()
     } else {
       // 只要error，cancel 所有请求
@@ -695,6 +695,7 @@ export class BaseDownloader extends BaseLoader {
         setTimeout(() => {
           this.retryAllDownloadRequest()
         })
+        this.message = e.message
         throw new Error('stopped')
       } else {
         throw e