+++ src/pkg.ts
@@ -1,6 +1,6 @@
 /** @format */
 
 export default {
-  "version": "0.2.4-beta.7",
+  "version": "0.2.4-beta.8",
   "name": "aliyun-pds-js-sdk"
 }