+++ src/utils/HttpUtil.ts
@@ -69,9 +69,7 @@ function isStopableError(e: Error): boolean {
 
 // 这些错误，重试多次后，暂停，下次可以断点续传
 function isNetworkError(e: Error): boolean {
-  return /Network Error|socket hang up|getaddrinfo ENOTFOUND|timeout |connect ECONNRESET|connect ETIMEDOUT|EPIPE/.test(
-    e.message,
-  )
+  return /Network Error|socket hang up|getaddrinfo ENOTFOUND|timeout | ECONNRESET| ETIMEDOUT|EPIPE/.test(e.message)
 }
 function isOssUrlExpired(e: AxiosError): boolean {
   return (